불필요 함수 삭제, 검색 버튼 클릭 시 샘플 데이터 전시
