+++ projects/kim/src/app/core/models/city.model.ts
@@ -1,4 +1,4 @@
 export interface City {
     title: string;
-    woeid: string;
+    woeid: number;
 }
