+++ projects/kim/src/app/core/services/city.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { WeatherCondition } from '../enums';
-import { City, Weather } from '../models';
+import { City } from '../models';
+import { ApiService } from './api.service';
 
 @Injectable({
   providedIn: 'root'
@@ -11,10 +11,9 @@ export class CityService {
   citySubject: BehaviorSubject<City> = new BehaviorSubject<City>(null);
   cityData$: Observable<City> = this.citySubject.asObservable();
 
-  weatherSubject: BehaviorSubject<Weather> = new BehaviorSubject<Weather>(null);
-  weatherData$: Observable<Weather> = this.weatherSubject.asObservable();
-
-  constructor() { }
+  constructor(
+    private apiService: ApiService,
+  ) { }
 
   getCityData(): Observable<City> {
     return this.cityData$;
@@ -23,24 +22,9 @@ export class CityService {
   setCityData(): void {
     const CITYDATA: City = {
       title: 'Seoul',
-      woeid: '1132599'
+      woeid: 1132599
     };
     this.citySubject.next(CITYDATA);
-  }
-
-  getWeatherData(): Observable<Weather> {
-    return this.weatherData$;
-  }
-
-  setWeatherData(): void {
-    const WEATHERDATA: Weather = {
-      condition: WeatherCondition.clear,
-      minTemp: 10,
-      maxTemp: 20,
-      Temp: 15,
-      updateDate: new Date()
-    };
-    this.weatherSubject.next(WEATHERDATA);
   }
 
 }
