+++ projects/kim/src/app/core/services/index.ts
@@ -1,4 +1,4 @@
-export * from './api.service';
-export * from './city.service';
-export * from './weather.service';
+// export * from './api.service';
+// export * from './city.service';
+// export * from './weather.service';
