+++ projects/kim/src/app/core/services/weather.service.ts
@@ -1,25 +1,35 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { City, Weather } from '../models';
+import { WeatherCondition } from '../enums';
+import { Weather } from '../models';
+import { ApiService } from './api.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class WeatherService {
-  citySubject: BehaviorSubject<City> = new BehaviorSubject<City>(null);
-  cityData$: Observable<City> = this.citySubject.asObservable();
 
   weatherSubject: BehaviorSubject<Weather> = new BehaviorSubject<Weather>(null);
   weatherData$: Observable<Weather> = this.weatherSubject.asObservable();
 
-  constructor() { }
+  constructor(
+    private apiService: ApiService,
+  ) { }
 
-  initCityData(): Observable<City> {
-    return this.cityData$;
-  }
-
-  initWeatherData(): Observable<Weather> {
+  getWeatherData(): Observable<Weather> {
     return this.weatherData$;
   }
 
+  setWeatherData(): void {
+    const WEATHERDATA: Weather = {
+      condition: WeatherCondition.clear,
+      minTemp: 10,
+      maxTemp: 20,
+      Temp: 15,
+      updateDate: new Date(),
+      location: 'Seoul',
+      locationId: 1132599
+    };
+    this.weatherSubject.next(WEATHERDATA);
+  }
 }
