+++ projects/kim/src/app/shared/components/city.search.component.html
@@ -2,4 +2,3 @@
     <input type="text" placeholder="Search City.." formControlName="city" />
     <button type='submit'>Search</button>
 </form>
-<button routerLink="/settings">Settings</button>
