+++ projects/kim/src/app/shared/components/city.search.component.ts
@@ -1,8 +1,6 @@
-import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { Observable } from 'rxjs';
-import { City, Weather } from '../../core/models';
-import { CityService } from '../../core/services';
+import { CityService } from '../../core/services/city.service';
 
 @Component({
     selector: 'app-city-search',
@@ -10,11 +8,6 @@ import { CityService } from '../../core/services';
 })
 export class CitySearchComponent implements OnInit {
     searchCityForm: FormGroup;
-    @Output() cityData = new EventEmitter<Observable<City>>();
-    @Output() weatherData = new EventEmitter<Observable<Weather>>();
-
-    cityData$: Observable<City>;
-    weatherData$: Observable<Weather>;
 
     constructor(
         private cityService: CityService,
@@ -26,12 +19,9 @@ export class CitySearchComponent implements OnInit {
     }
 
     ngOnInit(): void {
-        this.cityService.setCityData();
-        this.cityService.setWeatherData();
     }
 
     searchCity(): void {
-        this.cityData.emit(this.cityService.getCityData());
-        this.weatherData.emit(this.cityService.getWeatherData());
+        this.cityService.setCityData();
     }
 }
